requirements.txt
csignum-fast==1.0.2
mpmath==1.3.0
sympy==1.14.0
